requirements.txt
pytest
selene>=2.0.0rc2
python-dotenv==0.20.0
allure-pytest==2.9.45